Fix various issues with the sensor HAL
This commit fixes the following reported issues:
1) there are a lot of LOG spew
Silent Poll__setDelay and Poll__activiate LOG spew.
Changed "setDelay handle" to LOGD instead of of LOGE.

2) Gyro HAL reports a resolution of 0.28 rad/s, that is 16 deg/s. That seems wrong.
It appears the previous values we used was under the most sensitive settings, the new value is using an average setting instead.

3) The orientation sensor reports a name of "MPL Orientation (android deprecated format)". Please remove the "android deprecated format"
from the name.
Removed "android deprecated format"

4) MPL linear accel reports a max range of 10240.0 m/s^2. This seems wrong (that 1000+ Gs!!) Please rename to "MPL Linear Acceleration" (accel is not a word).
It appears that the dynamic sensor list does not recognize the accel id (or compass id) that were passed, causing the system to always read the initialized value. This is now fixed.
Renamed to MPL Linear Acceleration and MPL labels to be more humanly readable.

5) Fix for more accurate time stamps. This is a small subset of what we committed to Rocky's team and Kevin felt it was an easy change.
Instead of using system time stamp, the fix is to retrive timestamp directly from the driver when an IRQ is fired.

Change-Id: I8cfbd7cdecbd2e25613204ec1013e6079de97f37