Fix makefile's Android version check logic
Makefile's logic to determine if Android version was KitKat or
Jellybean would only work when PLATFORM_VERSION was defined as
a number, but PLATFORM_VERSION "can be an arbitrary string".
This broke the Invensense build when PLATFORM_VERSION changed
to KKWT.

Changed makefile to no longer check Android version, effectively
removing Jellybean support.

Change-Id: Ibbec6130cce4d91435d3a33b374aff871284100c
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>