Invensense: add opt-in line for the 60xx components
Add opt-in line so devices can sign up to use 60xx components.  Prior
to this fix, any device that wasn't a hammerhead, dory, or guppy
defaulted to parsing the 60xx folder's makefiles, regardless of whether
the device had an Invensense chip or not.  This was causing anthias builds
to break.

Change-Id: Ifa2c0fb96e985f1c81e1f5631bf53a0662aee978